+++ backend/main.py
@@ -7,7 +7,8 @@ from src.lib.visma import getVisma
 import json
 
 app = Flask(__name__, template_folder="../frontend")
-CORS(app, resources={r"/api/*": {"origins": "*"}}) #Initialize flask Cors (Cross Origin Resource Sharing) on only the /api/* URI endpoints
+# Initialize flask Cors (Cross Origin Resource Sharing) on only the /api/* URI endpoints
+CORS(app, resources={r"/api/*": {"origins": "*"}})
 
 
 @app.route('/')  # Index path
@@ -23,7 +24,7 @@ def serve_static(filename):
 
 @app.route('/api/friminutt', methods=['GET'])
 def getFriminutt():
-    
+
     def liste(friminutt):
         today = datetime.now()
         current_time = datetime.now().time()
@@ -33,7 +34,8 @@ def getFriminutt():
             res = [eval(i) for i in split_tider]
             test = datetime(today.year, today.month, today.day, res[0], res[1])
             if today < test:
-                current_datetime = datetime.combine(datetime.today(), current_time)
+                current_datetime = datetime.combine(
+                    datetime.today(), current_time)
                 time_difference = test - current_datetime
                 total_minutes_until_recess = time_difference.total_seconds() / 60
                 percentage_difference = (total_minutes_until_recess / 40) * 100
@@ -42,13 +44,13 @@ def getFriminutt():
                 inverted_percentage = 100 - percentage_difference
 
                 return friminutt[x], inverted_percentage
-            
 
     # Calculate the break times within the route handler
-    friminutt_start = ["9:5", "9:55", "10:55", "11:45", "13:00", "13:50", "14:40"]
+    friminutt_start = ["9:5", "9:55", "10:55",
+                       "11:45", "13:00", "13:50", "14:40"]
     friminutter = liste(friminutt_start)
     print(friminutter)
-    return {"friminutt": friminutter[1]}
+    return {"friminutt": "friminutter[1]"}
 
 
 @app.route('/api/joke', methods=['GET'])
@@ -70,4 +72,4 @@ def getWeather():
 if __name__ == '__main__':
 
     # starter app i debug mode som gjør at den reloader serveren on save
-    app.run(port=8080, debug=True, host="0.0.0.0")
+    app.run(port=5500, debug=True, host="0.0.0.0")
