--- backend/src/lib/.env
@@ -1,3 +1,0 @@
-feidenavn = 23F7CX9QM8amywaNGp_wnWdBQUFBQUJsSG1uSHM4cVNMcVlxUUlfMGt5WUFXdy04RjQ4Zm9aLVVMbGg1YjdqZnpfMjNQampFbXVuNldob1lnVVZkOENxb09zUHJfSFBKME9ZRnU2MW1JYTVqbTE2QktRPT0=
-feidepassord = Tc_LcnyY42bGC5jRNRtXTGdBQUFBQUJsSGxfY2ZuczlERUhGNWJNUnFKZnNHWFZ0T0VIa2F0Tm82Mld1ckFJNU42XzJrNWo4cGJvSG5YRkZiZV9UVlpRYUdTdjd6RXNXRlUtbDgxdUNVUFU5c0taYkdBPT0=
-SECRET = $hxu3I!URQX720iyLj84PFj
